+++ models/userdao.js
@@ -4,7 +4,7 @@ module.exports =
 {
   getAll()
   {
-    return DB.query('SELECT * FROM users')
+    return DB.accessor.query('SELECT * FROM users')
       .then((result) =>
       {
         return result;
@@ -16,7 +16,7 @@ module.exports =
   },
   getById(id)
   {
-    return DB.query('SELECT * FROM users WHERE id = $(userID)',{userID: id})
+    return DB.accessor.query('SELECT * FROM users WHERE id = $(userID)',{userID: id})
       .then((result) =>
       {
         return result;
@@ -28,8 +28,8 @@ module.exports =
   },
   newUser(pName, pEmail, pAlliance_id)
   {
-    DB.query('INSERT INTO users(name, email, alliance_id) VALUES ($(name), $(email), $(alliance_id))',{name: pName, email: pEmail, alliance_id: pAlliance_id});
-    return DB.query('SELECT * FROM users WHERE name=$(name) AND email=$(email) AND alliance_id=$(alliance_id)',{name: pName, email: pEmail, alliance_id: pAlliance_id})
+    DB.accessor.query('INSERT INTO users(name, email, alliance_id) VALUES ($(name), $(email), $(alliance_id))',{name: pName, email: pEmail, alliance_id: pAlliance_id});
+    return DB.accessor.query('SELECT * FROM users WHERE name=$(name) AND email=$(email) AND alliance_id=$(alliance_id)',{name: pName, email: pEmail, alliance_id: pAlliance_id})
     .then((result) =>
     {
       return result;
@@ -41,10 +41,10 @@ module.exports =
   },
   delUser(id)
   {
-    return DB.query('SELECT * FROM users WHERE id = $(userID)',{userID: id})
+    return DB.accessor.query('DELETE FROM users WHERE id = $(userID)',{userID: id}) //DB.accessor.query('SELECT * FROM users WHERE id = $(userID)',{userID: id})
     .then((result) =>
     {
-      DB.query('DELETE FROM users WHERE id = $(userID)',{userID: id});
+      //DB.accessor.query('DELETE FROM users WHERE id = $(userID)',{userID: id});
       return result;
     })
     .catch((error) =>
@@ -54,8 +54,8 @@ module.exports =
   },
   putUser(pId,pName,pEmail,pAlliance_id)
   {
-    DB.query('UPDATE users SET name=$(lName), email=$(lEmail), alliance_id=$(lAlliance_id) WHERE id=$(lId)',{lId: pId, lName: pName, lEmail: pEmail, lAlliance_id: pAlliance_id})
-    return DB.query('SELECT * FROM users')
+    DB.accessor.query('UPDATE users SET name=$(lName), email=$(lEmail), alliance_id=$(lAlliance_id) WHERE id=$(lId)',{lId: pId, lName: pName, lEmail: pEmail, lAlliance_id: pAlliance_id})
+    return DB.accessor.query('SELECT * FROM users WHERE id=$(lId)',{lId: pId})
     .then((result) =>
     {
       return result;
@@ -68,7 +68,7 @@ module.exports =
   //BONUS
   getCharacter(pId)
   {
-    return DB.query('SELECT * FROM characters WHERE user_id=$(lId)',{lId: pId})
+    return DB.accessor.query('SELECT * FROM characters WHERE user_id=$(lId)',{lId: pId})
     .then((result) =>
     {
       return result;
